requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.5
colorama==0.4.6
cryptography
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.2
pydantic_core==2.1.2
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
jinja2
psycopg2-binary
boto3
loguru
pyyaml
python-dotenv
twilio
Pillow
httpx # THIS IS USED FOR ASYNCHRONOUS , LIKE WHEN I WAS UPLOADING FILES TO S3
# tesserocr
pytesseract
scikit-image
imutils
opencv-python
numpy

# pip cache purge
# pip install -r requirements.txt